perf(generate): shrink parse/lex/production ids from `usize` to `u32`
`ParseStateId`, `LexStateId`, `ProductionInfoId`, and `ReservedWordSetId`
each held a `usize`, while the runtime stores state/field ids as `u16` (so
`u32` has plenty of headroom). Shrinking them saves space for every parse-table
entry, and a number of other data structures.

Reduces wall time by 2-5%, peak rss by 9-12%.

+++ crates/generate/src/grammars.rs
@@ -217,7 +217,7 @@ impl ProductionStep {
 // Extracted syntax grammar
 
 #[derive(Clone, Copy, Debug, Default, PartialEq, Eq, Hash, PartialOrd, Ord)]
-pub struct ReservedWordSetId(pub usize);
+pub struct ReservedWordSetId(pub u32);
 
 impl std::fmt::Display for ReservedWordSetId {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
